+++ projects/amdsmi/rocm_smi/src/rocm_smi_board_temp.cc
@@ -460,7 +460,8 @@ rsmi_status_t read_gpuboard_temp_metrics(const char* filename,
     return ErrnoToRsmiStatus(errno);
   }
   // Always create hex dump for debugging, using the number of bytes actually read
-  std::string hexDump = createHexDump(&metrics, file.gcount(), "GPU Board Temperature Metrics");
+  std::string hexDump =
+      createHexDump(&metrics, static_cast<size_t>(file.gcount()), "GPU Board Temperature Metrics");
   LOG_DEBUG(hexDump);
 
   if (file.gcount() != sizeof(metrics)) {
@@ -536,7 +537,8 @@ rsmi_status_t read_baseboard_temp_metrics(const char* filename,
     return ErrnoToRsmiStatus(errno);
   }
   // Always create hex dump for debugging, using the number of bytes actually read
-  std::string hexDump = createHexDump(&metrics, file.gcount(), "Baseboard Temperature Metrics");
+  std::string hexDump =
+      createHexDump(&metrics, static_cast<size_t>(file.gcount()), "Baseboard Temperature Metrics");
   LOG_DEBUG(hexDump);
 
   if (file.gcount() != sizeof(metrics)) {
