+++ projects/amdsmi/rocm_smi/src/rocm_smi_gpu_metrics.cc
@@ -100,7 +100,8 @@ constexpr bool is_apu_metrics_v30(const AMDGpuMetricsHeader_v1_t& header) {
 
 uint64_t actual_timestamp_in_secs() {
   using namespace std::chrono;
-  return duration_cast<seconds>(system_clock::now().time_since_epoch()).count();
+  return static_cast<uint64_t>(
+      duration_cast<seconds>(system_clock::now().time_since_epoch()).count());
 }
 
 auto timestamp_to_time_point(uint64_t timestamp_in_secs) {
